notebook-clone/notebook-clone/backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
llama-parse
openai
python-dotenv
tiktoken